+++ doc/CHANGELOG.md
@@ -5,9 +5,9 @@
 - GraphQL updates
   - Added `user` resolver and added `user` field on `history` type
   - Added `channels` resolver to list all channels for all adapters
+  - Added history_item` resolver to get a single history item by ID
 - Upgrade to `yetibot-dashboard 0.6.TODO`
-- Fix Docker Compose env var Slack example
-  [#738](https://github.com/yetibot/yetibot/pull/738) by @jcorrado
+- Switch to [#41](https://github.com/yetibot/yetibot.core/pull/41)
 
 ## 0.4.41 - 9/28/2018